dom/base/test/chrome/test_bug289714.xhtml

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=289714
-->
<window title="Mozilla Bug 289714"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=289714"
     target="_blank">Mozilla Bug 289714</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
  /** Test for Bug 289714 **/

  SimpleTest.waitForExplicitFinish();

  let xhr = new XMLHttpRequest();
  xhr.responseType = "document";
  xhr.open("GET", "data:text/xml,<xml");
  ok(xhr.channel !== undefined, "System XHRs should be privileged");
  xhr.onload = () => {
    ok(xhr.responseXML !== null, "System XHRs should yield <parsererrors>");
    SimpleTest.finish();
  };
  xhr.send();
  ]]></script>
</window>
